requirements.txt
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-WTF==1.0.0
WTForms-SQLAlchemy==0.3
WTForms~=3.0.1
Flask-Login==0.5.0
email-validator==1.1.3
